Remove hardcoded references to the registry gRPC API from the resolver packages. (#2340)
* Stop using the registry's Bundle proto type for resolution.

The resolver component does not care whether a given cache entry was
populated by use of the registry gRPC API, so it should ultimately be
possible to drop the "Bundle" field from cache entries altogether. For
now, catalog-operator's "step resolver" still depends on a Bundle
proto field in order to unpack inline (as opposed to image-based)
packages.


* Remove exported resolver cache entry creation funcs.

NewOperatorForBundle and NewOperatorFromV1Alpha1CSV couple the
resolver cache package to the registry gRPC API and the
operators/v1alpha1 CSV API, respectively. They're effectively used
only by the resolver package, so they can be moved there and
unexported. The remaining calls to NewOperatorFromV1Alpha1CSV can be
extracted from core resolver logic into a Source implementation that
is responsible for the conversion from ClusterServiceVersion to
cache.Snapshot entry, and the remaining call to NewOperatorFromBundle
is used only by the Source implementation based on the registry gRPC
API.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: b849eb79c6fbb1c445ed3e9945e1668f6066ea6e
