+++ staging/operator-lifecycle-manager/pkg/controller/operators/olm/labeler.go
@@ -1,8 +1,13 @@
 package olm
 
 import (
+	"fmt"
+	"strings"
+
+	"github.com/operator-framework/api/pkg/operators/v1alpha1"
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/controller/registry/resolver/cache"
 	"github.com/operator-framework/operator-registry/pkg/registry"
+	opregistry "github.com/operator-framework/operator-registry/pkg/registry"
 	extv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	extv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -13,9 +18,41 @@ const (
 	APILabelKeyPrefix = "olm.api."
 )
 
-type operatorSurface interface {
-	GetProvidedAPIs() cache.APISet
-	GetRequiredAPIs() cache.APISet
+type operatorSurface struct {
+	ProvidedAPIs cache.APISet
+	RequiredAPIs cache.APISet
+}
+
+func apiSurfaceOfCSV(csv *v1alpha1.ClusterServiceVersion) (*operatorSurface, error) {
+	surface := operatorSurface{
+		ProvidedAPIs: cache.EmptyAPISet(),
+		RequiredAPIs: cache.EmptyAPISet(),
+	}
+
+	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Owned {
+		parts := strings.SplitN(crdDef.Name, ".", 2)
+		if len(parts) < 2 {
+			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
+		}
+		surface.ProvidedAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
+	}
+	for _, api := range csv.Spec.APIServiceDefinitions.Owned {
+		surface.ProvidedAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
+	}
+
+	requiredAPIs := cache.EmptyAPISet()
+	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Required {
+		parts := strings.SplitN(crdDef.Name, ".", 2)
+		if len(parts) < 2 {
+			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
+		}
+		requiredAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
+	}
+	for _, api := range csv.Spec.APIServiceDefinitions.Required {
+		requiredAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
+	}
+
+	return &surface, nil
 }
 
 // LabelSetsFor returns API label sets for the given object.
@@ -35,14 +72,14 @@ func LabelSetsFor(obj interface{}) ([]labels.Set, error) {
 
 func labelSetsForOperatorSurface(surface operatorSurface) ([]labels.Set, error) {
 	labelSet := labels.Set{}
-	for key := range surface.GetProvidedAPIs().StripPlural() {
+	for key := range surface.ProvidedAPIs.StripPlural() {
 		hash, err := cache.APIKeyToGVKHash(key)
 		if err != nil {
 			return nil, err
 		}
 		labelSet[APILabelKeyPrefix+hash] = "provided"
 	}
-	for key := range surface.GetRequiredAPIs().StripPlural() {
+	for key := range surface.RequiredAPIs.StripPlural() {
 		hash, err := cache.APIKeyToGVKHash(key)
 		if err != nil {
 			return nil, err
