+++ staging/operator-lifecycle-manager/pkg/controller/registry/resolver/cache/operators_test.go
@@ -1,17 +1,10 @@
 package cache
 
 import (
-	"encoding/json"
 	"testing"
 
-	"github.com/blang/semver/v4"
 	"github.com/stretchr/testify/require"
-	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	opver "github.com/operator-framework/api/pkg/lib/version"
-	"github.com/operator-framework/api/pkg/operators/v1alpha1"
-	"github.com/operator-framework/operator-registry/pkg/api"
 	opregistry "github.com/operator-framework/operator-registry/pkg/registry"
 )
 
@@ -884,607 +877,4 @@ func TestOperatorSourceInfo_String(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestNewOperatorFromBundle(t *testing.T) {
-	version := opver.OperatorVersion{Version: semver.MustParse("0.1.0-abc")}
-	csv := v1alpha1.ClusterServiceVersion{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       v1alpha1.ClusterServiceVersionKind,
-			APIVersion: v1alpha1.GroupVersion,
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "testCSV",
-			Namespace: "testNamespace",
-		},
-		Spec: v1alpha1.ClusterServiceVersionSpec{
-			Replaces: "v1",
-			CustomResourceDefinitions: v1alpha1.CustomResourceDefinitions{
-				Owned:    []v1alpha1.CRDDescription{},
-				Required: []v1alpha1.CRDDescription{},
-			},
-			APIServiceDefinitions: v1alpha1.APIServiceDefinitions{
-				Owned:    []v1alpha1.APIServiceDescription{},
-				Required: []v1alpha1.APIServiceDescription{},
-			},
-			Version: version,
-		},
-	}
-
-	csvJson, err := json.Marshal(csv)
-	require.NoError(t, err)
-	bundleNoAPIs := &api.Bundle{
-		CsvName:     "testBundle",
-		PackageName: "testPackage",
-		ChannelName: "testChannel",
-		Version:     version.String(),
-		CsvJson:     string(csvJson),
-		Object:      []string{string(csvJson)},
-	}
-
-	csv.Spec.CustomResourceDefinitions.Owned = []v1alpha1.CRDDescription{
-		{
-			Name:    "owneds.crd.group.com",
-			Version: "v1",
-			Kind:    "OwnedCRD",
-		},
-	}
-	csv.Spec.CustomResourceDefinitions.Required = []v1alpha1.CRDDescription{
-		{
-			Name:    "requireds.crd.group.com",
-			Version: "v1",
-			Kind:    "RequiredCRD",
-		},
-	}
-	csv.Spec.APIServiceDefinitions.Owned = []v1alpha1.APIServiceDescription{
-		{
-			Name:    "ownedapis",
-			Group:   "apis.group.com",
-			Version: "v1",
-			Kind:    "OwnedAPI",
-		},
-	}
-	csv.Spec.APIServiceDefinitions.Required = []v1alpha1.APIServiceDescription{
-		{
-			Name:    "requiredapis",
-			Group:   "apis.group.com",
-			Version: "v1",
-			Kind:    "RequiredAPI",
-		},
-	}
-
-	csvJsonWithApis, err := json.Marshal(csv)
-	require.NoError(t, err)
-
-	crd := v1beta1.CustomResourceDefinition{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "CustomResourceDefinition",
-			APIVersion: "apiextensions.k8s.io/v1beta1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "owneds.crd.group.com",
-		},
-		Spec: v1beta1.CustomResourceDefinitionSpec{
-			Group: "crd.group.com",
-			Versions: []v1beta1.CustomResourceDefinitionVersion{
-				{
-					Name:    "v1",
-					Served:  true,
-					Storage: true,
-				},
-			},
-			Names: v1beta1.CustomResourceDefinitionNames{
-				Plural:   "owneds",
-				Singular: "owned",
-				Kind:     "OwnedCRD",
-				ListKind: "OwnedCRDList",
-			},
-		},
-	}
-	crdJson, err := json.Marshal(crd)
-	require.NoError(t, err)
-
-	bundleWithAPIs := &api.Bundle{
-		CsvName:     "testBundle",
-		PackageName: "testPackage",
-		ChannelName: "testChannel",
-		Version:     version.String(),
-		CsvJson:     string(csvJsonWithApis),
-		Object:      []string{string(csvJsonWithApis), string(crdJson)},
-		ProvidedApis: []*api.GroupVersionKind{
-			{
-				Group:   "crd.group.com",
-				Version: "v1",
-				Kind:    "OwnedCRD",
-				Plural:  "owneds",
-			},
-			{
-				Plural:  "ownedapis",
-				Group:   "apis.group.com",
-				Version: "v1",
-				Kind:    "OwnedAPI",
-			},
-		},
-		RequiredApis: []*api.GroupVersionKind{
-			{
-				Group:   "crd.group.com",
-				Version: "v1",
-				Kind:    "RequiredCRD",
-				Plural:  "requireds",
-			},
-			{
-				Plural:  "requiredapis",
-				Group:   "apis.group.com",
-				Version: "v1",
-				Kind:    "RequiredAPI",
-			},
-		},
-	}
-
-	bundleWithPropsAndDeps := &api.Bundle{
-		CsvName:     "testBundle",
-		PackageName: "testPackage",
-		ChannelName: "testChannel",
-		Version:     version.String(),
-		BundlePath:  "image",
-		Properties: []*api.Property{
-			{
-				Type:  "olm.gvk",
-				Value: "{\"group\":\"crd.group.com\",\"kind\":\"OwnedCRD\",\"version\":\"v1\"}",
-			},
-			{
-				Type:  "olm.gvk",
-				Value: "{\"group\":\"apis.group.com\",\"kind\":\"OwnedAPI\",\"version\":\"v1\"}",
-			},
-		},
-		Dependencies: []*api.Dependency{
-			{
-				Type:  "olm.gvk",
-				Value: "{\"group\":\"crd.group.com\",\"kind\":\"RequiredCRD\",\"version\":\"v1\"}",
-			},
-			{
-				Type:  "olm.gvk",
-				Value: "{\"group\":\"apis.group.com\",\"kind\":\"RequiredAPI\",\"version\":\"v1\"}",
-			},
-		},
-	}
-
-	bundleWithAPIsUnextracted := &api.Bundle{
-		CsvName:     "testBundle",
-		PackageName: "testPackage",
-		ChannelName: "testChannel",
-		CsvJson:     string(csvJsonWithApis),
-		Object:      []string{string(csvJsonWithApis), string(crdJson)},
-	}
-
-	type args struct {
-		bundle         *api.Bundle
-		sourceKey      SourceKey
-		replaces       string
-		defaultChannel string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *Operator
-		wantErr error
-	}{
-		{
-			name: "BundleNoAPIs",
-			args: args{
-				bundle:    bundleNoAPIs,
-				sourceKey: SourceKey{Name: "source", Namespace: "testNamespace"},
-			},
-			want: &Operator{
-				Name:         "testBundle",
-				Version:      &version.Version,
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: EmptyAPISet(),
-				Bundle:       bundleNoAPIs,
-				SourceInfo: &OperatorSourceInfo{
-					Package: "testPackage",
-					Channel: "testChannel",
-					Catalog: SourceKey{Name: "source", Namespace: "testNamespace"},
-				},
-			},
-		},
-		{
-			name: "BundleWithAPIs",
-			args: args{
-				bundle:    bundleWithAPIs,
-				sourceKey: SourceKey{Name: "source", Namespace: "testNamespace"},
-			},
-			want: &Operator{
-				Name:    "testBundle",
-				Version: &version.Version,
-				ProvidedAPIs: APISet{
-					opregistry.APIKey{
-						Group:   "crd.group.com",
-						Version: "v1",
-						Kind:    "OwnedCRD",
-						Plural:  "owneds",
-					}: struct{}{},
-					opregistry.APIKey{
-						Group:   "apis.group.com",
-						Version: "v1",
-						Kind:    "OwnedAPI",
-						Plural:  "ownedapis",
-					}: struct{}{},
-				},
-				RequiredAPIs: APISet{
-					opregistry.APIKey{
-						Group:   "crd.group.com",
-						Version: "v1",
-						Kind:    "RequiredCRD",
-						Plural:  "requireds",
-					}: struct{}{},
-					opregistry.APIKey{
-						Group:   "apis.group.com",
-						Version: "v1",
-						Kind:    "RequiredAPI",
-						Plural:  "requiredapis",
-					}: struct{}{},
-				},
-				Properties: []*api.Property{
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"crd.group.com\",\"kind\":\"OwnedCRD\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"apis.group.com\",\"kind\":\"OwnedAPI\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"crd.group.com\",\"kind\":\"RequiredCRD\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"apis.group.com\",\"kind\":\"RequiredAPI\",\"version\":\"v1\"}",
-					},
-				},
-				Bundle: bundleWithAPIs,
-				SourceInfo: &OperatorSourceInfo{
-					Package: "testPackage",
-					Channel: "testChannel",
-					Catalog: SourceKey{Name: "source", Namespace: "testNamespace"},
-				},
-			},
-		},
-		{
-			name: "BundleIgnoreCSV",
-			args: args{
-				bundle:    bundleWithAPIsUnextracted,
-				sourceKey: SourceKey{Name: "source", Namespace: "testNamespace"},
-			},
-			want: &Operator{
-				Name:         "testBundle",
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: EmptyAPISet(),
-				Bundle:       bundleWithAPIsUnextracted,
-				SourceInfo: &OperatorSourceInfo{
-					Package: "testPackage",
-					Channel: "testChannel",
-					Catalog: SourceKey{Name: "source", Namespace: "testNamespace"},
-				},
-			},
-		},
-		{
-			name: "BundleInDefaultChannel",
-			args: args{
-				bundle:         bundleNoAPIs,
-				sourceKey:      SourceKey{Name: "source", Namespace: "testNamespace"},
-				defaultChannel: "testChannel",
-			},
-			want: &Operator{
-				Name:         "testBundle",
-				Version:      &version.Version,
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: EmptyAPISet(),
-				Bundle:       bundleNoAPIs,
-				SourceInfo: &OperatorSourceInfo{
-					Package:        "testPackage",
-					Channel:        "testChannel",
-					Catalog:        SourceKey{Name: "source", Namespace: "testNamespace"},
-					DefaultChannel: true,
-				},
-			},
-		},
-		{
-			name: "BundleWithPropertiesAndDependencies",
-			args: args{
-				bundle:    bundleWithPropsAndDeps,
-				sourceKey: SourceKey{Name: "source", Namespace: "testNamespace"},
-			},
-			want: &Operator{
-				Name:         "testBundle",
-				Version:      &version.Version,
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: EmptyAPISet(),
-				Properties: []*api.Property{
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"crd.group.com\",\"kind\":\"OwnedCRD\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"apis.group.com\",\"kind\":\"OwnedAPI\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"crd.group.com\",\"kind\":\"RequiredCRD\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"apis.group.com\",\"kind\":\"RequiredAPI\",\"version\":\"v1\"}",
-					},
-				},
-				Bundle: bundleWithPropsAndDeps,
-				SourceInfo: &OperatorSourceInfo{
-					Package: "testPackage",
-					Channel: "testChannel",
-					Catalog: SourceKey{Name: "source", Namespace: "testNamespace"},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewOperatorFromBundle(tt.args.bundle, "", tt.args.sourceKey, tt.args.defaultChannel)
-			require.Equal(t, tt.wantErr, err)
-			requirePropertiesEqual(t, tt.want.Properties, got.Properties)
-			tt.want.Properties, got.Properties = nil, nil
-			require.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestNewOperatorFromCSV(t *testing.T) {
-	version := opver.OperatorVersion{Version: semver.MustParse("0.1.0-abc")}
-	type args struct {
-		csv *v1alpha1.ClusterServiceVersion
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *Operator
-		wantErr error
-	}{
-		{
-			name: "NoProvided/NoRequired",
-			args: args{
-				csv: &v1alpha1.ClusterServiceVersion{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "operator.v1",
-					},
-					Spec: v1alpha1.ClusterServiceVersionSpec{
-						Version: version,
-					},
-				},
-			},
-			want: &Operator{
-				Name:         "operator.v1",
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: EmptyAPISet(),
-				SourceInfo:   &ExistingOperator,
-				Version:      &version.Version,
-			},
-		},
-		{
-			name: "Provided/NoRequired",
-			args: args{
-				csv: &v1alpha1.ClusterServiceVersion{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "operator.v1",
-					},
-					Spec: v1alpha1.ClusterServiceVersionSpec{
-						Version: version,
-						CustomResourceDefinitions: v1alpha1.CustomResourceDefinitions{
-							Owned: []v1alpha1.CRDDescription{
-								{
-									Name:    "crdkinds.g",
-									Version: "v1",
-									Kind:    "CRDKind",
-								},
-							},
-						},
-						APIServiceDefinitions: v1alpha1.APIServiceDefinitions{
-							Owned: []v1alpha1.APIServiceDescription{
-								{
-									Name:    "apikinds",
-									Group:   "g",
-									Version: "v1",
-									Kind:    "APIKind",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: &Operator{
-				Name: "operator.v1",
-				ProvidedAPIs: map[opregistry.APIKey]struct{}{
-					{Group: "g", Version: "v1", Kind: "APIKind", Plural: "apikinds"}: {},
-					{Group: "g", Version: "v1", Kind: "CRDKind", Plural: "crdkinds"}: {},
-				},
-				Properties: []*api.Property{
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"g\",\"kind\":\"APIKind\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"g\",\"kind\":\"CRDKind\",\"version\":\"v1\"}",
-					},
-				},
-				RequiredAPIs: EmptyAPISet(),
-				SourceInfo:   &ExistingOperator,
-				Version:      &version.Version,
-			},
-		},
-		{
-			name: "NoProvided/Required",
-			args: args{
-				csv: &v1alpha1.ClusterServiceVersion{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "operator.v1",
-					},
-					Spec: v1alpha1.ClusterServiceVersionSpec{
-						Version: version,
-						CustomResourceDefinitions: v1alpha1.CustomResourceDefinitions{
-							Required: []v1alpha1.CRDDescription{
-								{
-									Name:    "crdkinds.g",
-									Version: "v1",
-									Kind:    "CRDKind",
-								},
-							},
-						},
-						APIServiceDefinitions: v1alpha1.APIServiceDefinitions{
-							Required: []v1alpha1.APIServiceDescription{
-								{
-									Name:    "apikinds",
-									Group:   "g",
-									Version: "v1",
-									Kind:    "APIKind",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: &Operator{
-				Name:         "operator.v1",
-				ProvidedAPIs: EmptyAPISet(),
-				RequiredAPIs: map[opregistry.APIKey]struct{}{
-					{Group: "g", Version: "v1", Kind: "APIKind", Plural: "apikinds"}: {},
-					{Group: "g", Version: "v1", Kind: "CRDKind", Plural: "crdkinds"}: {},
-				},
-				Properties: []*api.Property{
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"g\",\"kind\":\"APIKind\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"g\",\"kind\":\"CRDKind\",\"version\":\"v1\"}",
-					},
-				},
-				SourceInfo: &ExistingOperator,
-				Version:    &version.Version,
-			},
-		},
-		{
-			name: "Provided/Required",
-			args: args{
-				csv: &v1alpha1.ClusterServiceVersion{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "operator.v1",
-					},
-					Spec: v1alpha1.ClusterServiceVersionSpec{
-						Version: version,
-						CustomResourceDefinitions: v1alpha1.CustomResourceDefinitions{
-							Owned: []v1alpha1.CRDDescription{
-								{
-									Name:    "crdownedkinds.g",
-									Version: "v1",
-									Kind:    "CRDOwnedKind",
-								},
-							},
-							Required: []v1alpha1.CRDDescription{
-								{
-									Name:    "crdreqkinds.g2",
-									Version: "v1",
-									Kind:    "CRDReqKind",
-								},
-							},
-						},
-						APIServiceDefinitions: v1alpha1.APIServiceDefinitions{
-							Owned: []v1alpha1.APIServiceDescription{
-								{
-									Name:    "apiownedkinds",
-									Group:   "g",
-									Version: "v1",
-									Kind:    "APIOwnedKind",
-								},
-							},
-							Required: []v1alpha1.APIServiceDescription{
-								{
-									Name:    "apireqkinds",
-									Group:   "g2",
-									Version: "v1",
-									Kind:    "APIReqKind",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: &Operator{
-				Name: "operator.v1",
-				ProvidedAPIs: map[opregistry.APIKey]struct{}{
-					{Group: "g", Version: "v1", Kind: "APIOwnedKind", Plural: "apiownedkinds"}: {},
-					{Group: "g", Version: "v1", Kind: "CRDOwnedKind", Plural: "crdownedkinds"}: {},
-				},
-				RequiredAPIs: map[opregistry.APIKey]struct{}{
-					{Group: "g2", Version: "v1", Kind: "APIReqKind", Plural: "apireqkinds"}: {},
-					{Group: "g2", Version: "v1", Kind: "CRDReqKind", Plural: "crdreqkinds"}: {},
-				},
-				Properties: []*api.Property{
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"g\",\"kind\":\"APIOwnedKind\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk",
-						Value: "{\"group\":\"g\",\"kind\":\"CRDOwnedKind\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"g2\",\"kind\":\"APIReqKind\",\"version\":\"v1\"}",
-					},
-					{
-						Type:  "olm.gvk.required",
-						Value: "{\"group\":\"g2\",\"kind\":\"CRDReqKind\",\"version\":\"v1\"}",
-					},
-				},
-				SourceInfo: &ExistingOperator,
-				Version:    &version.Version,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewOperatorFromV1Alpha1CSV(tt.args.csv)
-			require.Equal(t, tt.wantErr, err)
-			requirePropertiesEqual(t, tt.want.Properties, got.Properties)
-			tt.want.Properties, got.Properties = nil, nil
-			require.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func requirePropertiesEqual(t *testing.T, a, b []*api.Property) {
-	type Property struct {
-		Type  string
-		Value interface{}
-	}
-	nice := func(in *api.Property) Property {
-		var i interface{}
-		if err := json.Unmarshal([]byte(in.Value), &i); err != nil {
-			t.Fatalf("property value %q could not be unmarshaled as json: %s", in.Value, err)
-		}
-		return Property{
-			Type:  in.Type,
-			Value: i,
-		}
-	}
-	var l, r []Property
-	for _, p := range a {
-		l = append(l, nice(p))
-	}
-	for _, p := range b {
-		r = append(r, nice(p))
-	}
-	require.ElementsMatch(t, l, r)
 }
