+++ staging/operator-lifecycle-manager/pkg/controller/registry/resolver/cache/predicates.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/blang/semver/v4"
 
-	"github.com/operator-framework/operator-registry/pkg/api"
 	opregistry "github.com/operator-framework/operator-registry/pkg/registry"
 )
 
@@ -41,10 +40,10 @@ func (ch channelPredicate) Test(o *Operator) bool {
 	if string(ch) == "" {
 		return true
 	}
-	if o.Bundle == nil {
-		return false
+	if si := o.SourceInfo; si != nil {
+		return si.Channel == string(ch)
 	}
-	return o.Bundle.ChannelName == string(ch)
+	return false
 }
 
 func (ch channelPredicate) String() string {
@@ -322,67 +321,11 @@ func (c countingPredicate) Test(o *Operator) bool {
 	}
 	return false
 }
+
 func (c countingPredicate) String() string {
 	return c.p.String()
 }
+
 func CountingPredicate(p OperatorPredicate, n *int) OperatorPredicate {
 	return countingPredicate{p: p, n: n}
-}
-
-func PredicateForProperty(property *api.Property) (OperatorPredicate, error) {
-	if property == nil {
-		return nil, nil
-	}
-	p, ok := predicates[property.Type]
-	if !ok {
-		return nil, nil
-	}
-	return p(property.Value)
-}
-
-var predicates = map[string]func(string) (OperatorPredicate, error){
-	"olm.gvk.required":     predicateForRequiredGVKProperty,
-	"olm.package.required": predicateForRequiredPackageProperty,
-	"olm.label.required":   predicateForRequiredLabelProperty,
-}
-
-func predicateForRequiredGVKProperty(value string) (OperatorPredicate, error) {
-	var gvk struct {
-		Group   string `json:"group"`
-		Version string `json:"version"`
-		Kind    string `json:"kind"`
-	}
-	if err := json.Unmarshal([]byte(value), &gvk); err != nil {
-		return nil, err
-	}
-	return ProvidingAPIPredicate(opregistry.APIKey{
-		Group:   gvk.Group,
-		Version: gvk.Version,
-		Kind:    gvk.Kind,
-	}), nil
-}
-
-func predicateForRequiredPackageProperty(value string) (OperatorPredicate, error) {
-	var pkg struct {
-		PackageName  string `json:"packageName"`
-		VersionRange string `json:"versionRange"`
-	}
-	if err := json.Unmarshal([]byte(value), &pkg); err != nil {
-		return nil, err
-	}
-	ver, err := semver.ParseRange(pkg.VersionRange)
-	if err != nil {
-		return nil, err
-	}
-	return And(PkgPredicate(pkg.PackageName), VersionInRangePredicate(ver, pkg.VersionRange)), nil
-}
-
-func predicateForRequiredLabelProperty(value string) (OperatorPredicate, error) {
-	var label struct {
-		Label string `json:"label"`
-	}
-	if err := json.Unmarshal([]byte(value), &label); err != nil {
-		return nil, err
-	}
-	return LabelPredicate(label.Label), nil
 }
