+++ staging/operator-lifecycle-manager/pkg/controller/registry/resolver/resolver.go
@@ -76,7 +76,7 @@ func (r *SatResolver) SolveOperators(namespaces []string, csvs []*v1alpha1.Clust
 		var current *cache.Operator
 		for _, csv := range csvs {
 			if csv.Name == sub.Status.InstalledCSV {
-				op, err := cache.NewOperatorFromV1Alpha1CSV(csv)
+				op, err := newOperatorFromV1Alpha1CSV(csv)
 				if err != nil {
 					return nil, err
 				}
@@ -181,7 +181,7 @@ func (r *SatResolver) getSubscriptionInstallables(sub *v1alpha1.Subscription, cu
 		Namespace: sub.Spec.CatalogSourceNamespace,
 	}
 
-	var bundles []*cache.Operator
+	var entries []*cache.Operator
 	{
 		var nall, npkg, nch, ncsv int
 
@@ -200,7 +200,7 @@ func (r *SatResolver) getSubscriptionInstallables(sub *v1alpha1.Subscription, cu
 			cache.CountingPredicate(cache.ChannelPredicate(sub.Spec.Channel), &nch),
 			cache.CountingPredicate(csvPredicate, &ncsv),
 		))
-		bundles = namespacedCache.Catalog(catalog).Find(cachePredicates...)
+		entries = namespacedCache.Catalog(catalog).Find(cachePredicates...)
 
 		var si solver.Installable
 		switch {
@@ -220,36 +220,40 @@ func (r *SatResolver) getSubscriptionInstallables(sub *v1alpha1.Subscription, cu
 		}
 	}
 
-	// bundles in the default channel appear first, then lexicographically order by channel name
-	sort.SliceStable(bundles, func(i, j int) bool {
+	// entries in the default channel appear first, then lexicographically order by channel name
+	sort.SliceStable(entries, func(i, j int) bool {
 		var idef bool
-		if isrc := bundles[i].SourceInfo; isrc != nil {
+		var ichan string
+		if isrc := entries[i].SourceInfo; isrc != nil {
 			idef = isrc.DefaultChannel
+			ichan = isrc.Channel
 		}
 		var jdef bool
-		if jsrc := bundles[j].SourceInfo; jsrc != nil {
+		var jchan string
+		if jsrc := entries[j].SourceInfo; jsrc != nil {
 			jdef = jsrc.DefaultChannel
+			jchan = jsrc.Channel
 		}
 		if idef == jdef {
-			return bundles[i].Bundle.ChannelName < bundles[j].Bundle.ChannelName
+			return ichan < jchan
 		}
 		return idef
 	})
 
 	var sortedBundles []*cache.Operator
 	lastChannel, lastIndex := "", 0
-	for i := 0; i <= len(bundles); i++ {
-		if i != len(bundles) && bundles[i].Bundle.ChannelName == lastChannel {
+	for i := 0; i <= len(entries); i++ {
+		if i != len(entries) && entries[i].Channel() == lastChannel {
 			continue
 		}
-		channel, err := sortChannel(bundles[lastIndex:i])
+		channel, err := sortChannel(entries[lastIndex:i])
 		if err != nil {
 			return nil, err
 		}
 		sortedBundles = append(sortedBundles, channel...)
 
-		if i != len(bundles) {
-			lastChannel = bundles[i].Bundle.ChannelName
+		if i != len(entries) {
+			lastChannel = entries[i].Channel()
 			lastIndex = i
 		}
 	}
@@ -333,7 +337,7 @@ func (r *SatResolver) getBundleInstallables(preferredNamespace string, bundleSta
 
 		visited[bundle] = &bundleInstallable
 
-		dependencyPredicates, err := bundle.DependencyPredicates()
+		dependencyPredicates, err := DependencyPredicates(bundle.Properties)
 		if err != nil {
 			errs = append(errs, err)
 			continue
@@ -470,7 +474,7 @@ func (r *SatResolver) newSnapshotForNamespace(namespace string, subs []*v1alpha1
 	var csvsMissingProperties []*v1alpha1.ClusterServiceVersion
 	standaloneOperators := make([]*cache.Operator, 0)
 	for _, csv := range csvs {
-		op, err := cache.NewOperatorFromV1Alpha1CSV(csv)
+		op, err := newOperatorFromV1Alpha1CSV(csv)
 		if err != nil {
 			return nil, err
 		}
@@ -727,4 +731,173 @@ func sortChannel(bundles []*cache.Operator) ([]*cache.Operator, error) {
 
 	// TODO: do we care if the channel doesn't include every bundle in the input?
 	return chains[0], nil
+}
+
+func DependencyPredicates(properties []*api.Property) ([]cache.OperatorPredicate, error) {
+	var predicates []cache.OperatorPredicate
+	for _, property := range properties {
+		predicate, err := predicateForProperty(property)
+		if err != nil {
+			return nil, err
+		}
+		if predicate == nil {
+			continue
+		}
+		predicates = append(predicates, predicate)
+	}
+	return predicates, nil
+}
+
+func predicateForProperty(property *api.Property) (cache.OperatorPredicate, error) {
+	if property == nil {
+		return nil, nil
+	}
+	p, ok := predicates[property.Type]
+	if !ok {
+		return nil, nil
+	}
+	return p(property.Value)
+}
+
+var predicates = map[string]func(string) (cache.OperatorPredicate, error){
+	"olm.gvk.required":     predicateForRequiredGVKProperty,
+	"olm.package.required": predicateForRequiredPackageProperty,
+	"olm.label.required":   predicateForRequiredLabelProperty,
+}
+
+func predicateForRequiredGVKProperty(value string) (cache.OperatorPredicate, error) {
+	var gvk struct {
+		Group   string `json:"group"`
+		Version string `json:"version"`
+		Kind    string `json:"kind"`
+	}
+	if err := json.Unmarshal([]byte(value), &gvk); err != nil {
+		return nil, err
+	}
+	return cache.ProvidingAPIPredicate(opregistry.APIKey{
+		Group:   gvk.Group,
+		Version: gvk.Version,
+		Kind:    gvk.Kind,
+	}), nil
+}
+
+func predicateForRequiredPackageProperty(value string) (cache.OperatorPredicate, error) {
+	var pkg struct {
+		PackageName  string `json:"packageName"`
+		VersionRange string `json:"versionRange"`
+	}
+	if err := json.Unmarshal([]byte(value), &pkg); err != nil {
+		return nil, err
+	}
+	ver, err := semver.ParseRange(pkg.VersionRange)
+	if err != nil {
+		return nil, err
+	}
+	return cache.And(cache.PkgPredicate(pkg.PackageName), cache.VersionInRangePredicate(ver, pkg.VersionRange)), nil
+}
+
+func predicateForRequiredLabelProperty(value string) (cache.OperatorPredicate, error) {
+	var label struct {
+		Label string `json:"label"`
+	}
+	if err := json.Unmarshal([]byte(value), &label); err != nil {
+		return nil, err
+	}
+	return cache.LabelPredicate(label.Label), nil
+}
+
+func newOperatorFromV1Alpha1CSV(csv *v1alpha1.ClusterServiceVersion) (*cache.Operator, error) {
+	providedAPIs := cache.EmptyAPISet()
+	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Owned {
+		parts := strings.SplitN(crdDef.Name, ".", 2)
+		if len(parts) < 2 {
+			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
+		}
+		providedAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
+	}
+	for _, api := range csv.Spec.APIServiceDefinitions.Owned {
+		providedAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
+	}
+
+	requiredAPIs := cache.EmptyAPISet()
+	for _, crdDef := range csv.Spec.CustomResourceDefinitions.Required {
+		parts := strings.SplitN(crdDef.Name, ".", 2)
+		if len(parts) < 2 {
+			return nil, fmt.Errorf("error parsing crd name: %s", crdDef.Name)
+		}
+		requiredAPIs[opregistry.APIKey{Plural: parts[0], Group: parts[1], Version: crdDef.Version, Kind: crdDef.Kind}] = struct{}{}
+	}
+	for _, api := range csv.Spec.APIServiceDefinitions.Required {
+		requiredAPIs[opregistry.APIKey{Group: api.Group, Version: api.Version, Kind: api.Kind, Plural: api.Name}] = struct{}{}
+	}
+
+	properties, err := providedAPIsToProperties(providedAPIs)
+	if err != nil {
+		return nil, err
+	}
+	dependencies, err := requiredAPIsToProperties(requiredAPIs)
+	if err != nil {
+		return nil, err
+	}
+	properties = append(properties, dependencies...)
+
+	return &cache.Operator{
+		Name:         csv.GetName(),
+		Version:      &csv.Spec.Version.Version,
+		ProvidedAPIs: providedAPIs,
+		RequiredAPIs: requiredAPIs,
+		SourceInfo:   &cache.ExistingOperator,
+		Properties:   properties,
+	}, nil
+}
+
+func providedAPIsToProperties(apis cache.APISet) (out []*api.Property, err error) {
+	out = make([]*api.Property, 0)
+	for a := range apis {
+		val, err := json.Marshal(opregistry.GVKProperty{
+			Group:   a.Group,
+			Version: a.Version,
+			Kind:    a.Kind,
+		})
+		if err != nil {
+			panic(err)
+		}
+		out = append(out, &api.Property{
+			Type:  opregistry.GVKType,
+			Value: string(val),
+		})
+	}
+	if len(out) > 0 {
+		return
+	}
+	return nil, nil
+}
+
+func requiredAPIsToProperties(apis cache.APISet) (out []*api.Property, err error) {
+	if len(apis) == 0 {
+		return
+	}
+	out = make([]*api.Property, 0)
+	for a := range apis {
+		val, err := json.Marshal(struct {
+			Group   string `json:"group"`
+			Version string `json:"version"`
+			Kind    string `json:"kind"`
+		}{
+			Group:   a.Group,
+			Version: a.Version,
+			Kind:    a.Kind,
+		})
+		if err != nil {
+			return nil, err
+		}
+		out = append(out, &api.Property{
+			Type:  "olm.gvk.required",
+			Value: string(val),
+		})
+	}
+	if len(out) > 0 {
+		return
+	}
+	return nil, nil
 }
